pyproject.toml
[project]
name = "aio-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram==3.15.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "psycopg2-binary==2.9.10",
    "pyyaml>=6.0.3",
    "sqlalchemy==2.0.37",
    "sqlalchemy-orm>=1.2.10",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]


[tool.ruff]
line-length = 79
indent-width = 4
exclude = [
    "alembic",
    "migration",
    "migrations"
]


[tool.ruff.lint]
# Extend the default rule selection to include import sorting (I rules)
extend-select = ["I"]
fixable = ["I"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 79

[tool.isort]
line_length = 79
known_local_folder = []
